Add my GPG key and repository to the image
This is done because some AUR packages require other AUR dependencies.
makepkg just fails when that is the case. To solve this, for the
moment, I'm manually building those AUR dependencies, and uploading
them to my repo. Which means that by configuring my repo in this base
image, then the other builder images that build packages that have AUR
dependencies, will be able to find that AUR dependency as an
"official" dependency on my repo.